fix(global): list
